Split different cross correlation classes
