--- openquake/hazardlib/cross_correl_between.py
@@ -1,139 +1,0 @@
-# -*- coding: utf-8 -*-
-# vim: tabstop=4 shiftwidth=4 softtabstop=4
-#
-# Copyright (C) 2021, GEM Foundation
-#
-# OpenQuake is free software: you can redistribute it and/or modify it
-# under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OpenQuake is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with OpenQuake.  If not, see <http://www.gnu.org/licenses/>.
-
-import numpy as np
-import numpy.matlib
-from scipy import stats
-from abc import ABC, abstractmethod
-from openquake.hazardlib.imt import IMT
-
-
-class CrossCorrelationBetween(ABC):
-    def __init__(self, trunclevel=None):
-        self.trunclevel = trunclevel
-        if self.trunclevel is None:
-            self.distribution = stats.norm()
-        elif self.trunclevel == 0:
-            self.distribution = None
-        else:
-            self.distribution = stats.truncnorm(-trunclevel, trunclevel)
-
-    @abstractmethod
-    def get_correlation(self, from_imt: IMT, to_imt: IMT) -> float:
-        """
-        :param from_imt:
-            An intensity measure type
-        :param to_imt:
-            An intensity measure type
-        :return: a scalar
-        """
-    @abstractmethod
-    def get_inter_eps(self, imts, num_events):
-        pass
-
-
-class GodaAtkinson2009(CrossCorrelationBetween):
-    """
-    Implements the correlation model of Goda and Atkinson published in 2009
-    https://pubs.geoscienceworld.org/ssa/bssa/article-abstract/99/5/3003/342221/Probabilistic-Characterization-of-Spatially
-    """
-    cache = {}  # periods -> correlation matrix
-
-    def get_correlation(self, from_imt: IMT, to_imt: IMT) -> float:
-        """
-        :returns: a scalar in the range 0..1
-        """
-        if from_imt == to_imt:
-            return 1
-
-        T1 = from_imt.period or 0.05  # for PGA
-        T2 = to_imt.period or 0.05  # for PGA
-
-        Tmin = min(T1, T2)
-        Tmax = max(T1, T2)
-        ITmin = 1. if Tmin < 0.25 else 0.
-
-        theta1 = 1.374
-        theta2 = 5.586
-        theta3 = 0.728
-
-        angle = np.pi/2. - (theta1 + theta2 * ITmin * (Tmin / Tmax) ** theta3 *
-                            np.log10(Tmin / 0.25)) * np.log10(Tmax / Tmin)
-        delta = 1 + np.cos(-1.5 * np.log10(Tmin / Tmax))
-        return (1. - np.cos(angle) + delta) / 3.
-
-    def get_inter_eps(self, imts, num_events):
-        """
-        :param imts: a list of M intensity measure types
-        :param num_events: the number of events to consider (E)
-        :returns: a correlated matrix of epsilons of shape (M, E)
-
-        NB: the user must specify the random seed first
-        """
-        corma = self._get_correlation_matrix(imts)
-        return numpy.random.multivariate_normal(
-            numpy.zeros(len(imts)), corma, num_events)
-
-    def _get_correlation_matrix(self, imts):
-        # cached on the periods
-        periods = tuple(imt.period for imt in imts)
-        try:
-            return self.cache[periods]
-        except KeyError:
-            self.cache[periods] = corma = np.zeros((len(imts), len(imts)))
-        for i, imi in enumerate(imts):
-            for j, imj in enumerate(imts):
-                corma[i, j] = self.get_correlation(imi, imj)
-        return corma
-
-
-class NoCrossCorrelation(CrossCorrelationBetween):
-    """
-    Used when there is no cross correlation
-    """
-    def get_correlation(self, from_imt, to_imt):
-        return from_imt == to_imt
-
-    def get_inter_eps(self, imts, num_events):
-        """
-        :param imts: a list of M intensity measure types
-        :param num_events: the number of events to consider (E)
-        :returns: an uncorrelated matrix of epsilons of shape (M, E)
-
-        NB: the user must specify the random seed first
-        """
-        return numpy.array([self.distribution.rvs(num_events) for imt in imts])
-
-
-class FullCrossCorrelation(CrossCorrelationBetween):
-    """
-    Used when there is full cross correlation, i.e. same epsilons for all IMTs
-    """
-    def get_correlation(self, from_imt, to_imt):
-        return 1.
-
-    def get_inter_eps(self, imts, num_events):
-        """
-        :param imts: a list of M intensity measure types
-        :param num_events: the number of events to consider (E)
-        :returns: an uncorrelated matrix of epsilons of shape (M, E)
-
-        NB: the user must specify the random seed first
-        """
-        eps = self.distribution.rvs(num_events)
-        return numpy.array([eps for imt in imts])
